put loadJSON in login.dart
